really fixed microstepping

+++ Adafruit_STSPIN220.h
@@ -59,23 +59,23 @@ typedef enum {
  */
 class Adafruit_STSPIN220 {
  public:
-  Adafruit_STSPIN220(int number_of_steps, int step_pin, int dir_pin);
-  Adafruit_STSPIN220(int number_of_steps, int step_pin, int dir_pin,
-                     int mode1_pin, int mode2_pin, int en_fault_pin = -1,
-                     int stby_reset_pin = -1);
+  Adafruit_STSPIN220(int16_t number_of_steps, int16_t step_pin, int16_t dir_pin);
+  Adafruit_STSPIN220(int16_t number_of_steps, int16_t step_pin, int16_t dir_pin,
+                     int16_t mode1_pin, int16_t mode2_pin, int16_t en_fault_pin = -1,
+                     int16_t stby_reset_pin = -1);
 
   // Arduino Stepper library compatible interface
-  void setSpeed(long whatSpeed);
-  void step(int steps_to_move);
-  int version(void);
+  void setSpeed(int32_t whatSpeed);
+  void step(int32_t steps_to_move);
+  int16_t version(void);
 
   // Extended STSPIN220 specific functionality
   bool setStepMode(stspin220_step_mode_t mode);
   stspin220_step_mode_t getStepMode();
-  int microstepsPerStep();
+  int16_t microstepsPerStep();
 
   void singleStep();
-  void stepBlocking(int steps, unsigned long delay_us = 1000);
+  void stepBlocking(int32_t steps, unsigned long delay_us = 1000);
 
   void enable(bool state = true);
   bool isEnabled();
@@ -89,18 +89,18 @@ class Adafruit_STSPIN220 {
 
  private:
   // Arduino Stepper library compatible members
-  unsigned long _step_delay;     ///< Delay between steps in microseconds
-  int _number_of_steps;          ///< Total steps per revolution
-  int _step_number;              ///< Current position in steps
-  unsigned long _last_step_time; ///< Last step timestamp
+  uint32_t _step_delay;          ///< Delay between steps in microseconds
+  int16_t _number_of_steps;      ///< Total steps per revolution
+  int32_t _step_number;          ///< Current position in steps
+  uint32_t _last_step_time;      ///< Last step timestamp
 
   // STSPIN220 specific pins
-  int _step_pin;       ///< STEP/MODE3 pin (step clock)
-  int _dir_pin;        ///< DIR/MODE4 pin (direction)
-  int _mode1_pin;      ///< MODE1 pin (-1 if not used)
-  int _mode2_pin;      ///< MODE2 pin (-1 if not used)
-  int _en_fault_pin;   ///< EN/FAULT pin (-1 if not used)
-  int _stby_reset_pin; ///< STBY/RESET pin (-1 if not used)
+  int16_t _step_pin;       ///< STEP/MODE3 pin (step clock)
+  int16_t _dir_pin;        ///< DIR/MODE4 pin (direction)
+  int16_t _mode1_pin;      ///< MODE1 pin (-1 if not used)
+  int16_t _mode2_pin;      ///< MODE2 pin (-1 if not used)
+  int16_t _en_fault_pin;   ///< EN/FAULT pin (-1 if not used)
+  int16_t _stby_reset_pin; ///< STBY/RESET pin (-1 if not used)
 
   // STSPIN220 state
   stspin220_step_mode_t _step_mode; ///< Current microstepping mode
